chore(editor): split StatusTray.tsx into one-component-per-file (#439)
Extracts the six sub-components from the 845-line StatusTray.tsx into
dedicated files (EditorToolsGroup, CompletionsGroup, CommentsGroup,
ActionsGroup, SessionGroup, HelpGroup). Reduces the orchestrator to
~195 lines that compose the six groups. No behaviour changes.
